requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
colorama==0.4.6
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
setuptools==68.2.2
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
wheel==0.41.2
